amplify/agent/runtime/requirements.txt
bedrock-agentcore
strands-agents[bidi,otel]
aws-opentelemetry-distro
